Slightly refactored main app page and commented it
