+++ src/components/AppEdit.tsx
@@ -8,29 +8,22 @@ import useWindowWidth from '../hooks/useWindow';
 import ActionButtons from './ActionButtons';
 import DataGridWrapper from './DataGridWrapper';
 import type { ColumnConfig } from '../interfaces/ColumnData';
-
-// const customWidths: Record<string, string> = {
-//   When: "10%",
-//   EditedBy: "10%",
-//   Action: "10%",
-//   WhoWasEdited: "20%",
-//   Column: "10%",
-//   ChangedFrom: "20%",
-//   ChangedTo: "20%"
-// };
+import { editFiles, datasetLabels } from "../config/AppConfig";
+import { 
+  handleHomePage,
+  handleData,
+  handleEditHistory
+} from "../utils/gridHelper";
 
 export default function App() {
-  const editFiles: Record<string, string> = {
-    csprofs: "edit-history.csv",
-    students: "students-edit-history.csv",
-  };
-
+  // get the dataset we're in from the url
   const base = import.meta.env.BASE_URL;
   const dataset = window.location.pathname
     .replace(base, "")
     .split("/")
     .filter(Boolean)[0];
 
+  // create portal div if it doesn't exist to render modals into
   useEffect(() => {
     const portalDiv = document.getElementById('portal');
 
@@ -50,6 +43,7 @@ export default function App() {
   const [columnFilters, setColumnFilters] = useState<Record<string, string>>({});
 
   useEffect(() => {
+    // gather data from csv and yaml, and apply any filters from the url
     const fetchData = async () => {
       try {
         const { gridColumns, parsedData, columnSchema } = await fetchCsvData(
@@ -83,6 +77,7 @@ export default function App() {
     fetchData();
   }, [gridWidth]);
  
+  // re-apply filters and sorting whenever relevant state changes
   useEffect(() => {
     if (columns.length === 0 || data.length === 0) return;
 
@@ -108,6 +103,7 @@ export default function App() {
     setFilteredData(filtered);
   }, [columnFilters, data, columns]);
 
+  // update url when filters change
   useEffect(() => {
     if (columns.length === 0) return;
 
@@ -122,31 +118,16 @@ export default function App() {
     window.history.replaceState(null, '', '?' + params.toString());
   }, [columnFilters, columns]);
 
-  const handleHomePage = () => {
-    window.location.href = `${base}`;
-  }
-
-  const handleData = () => {
-    window.location.href = `${base}${dataset}`;
-  }
-
-  const handleEditHistory = () => {
-    window.location.href = `${base}${dataset}/history`;
-  };
-
-  const datasetLabels: Record<string, string> = {
-    csprofs: "CS Professors",
-    students: "Students",
-  };
-
+  // get the label for the current dataset to show in the UI
   const datasetLabel = datasetLabels[dataset];
 
-  return (
+  // render the app with the action buttons at the top and the data grid taking up the rest of the space
+  return ( 
     <div className="App" style={{ display: "flex", flexDirection: "column", height: "100vh" }}>
       <ActionButtons
-        handleHomePage={handleHomePage}
-        handleData={handleData}
-        handleEditHistory={handleEditHistory}
+        handleHomePage={() => handleHomePage(base)}
+        handleData={() => handleData(base, dataset)}
+        handleEditHistory={() => handleEditHistory(base, dataset)}
         datasetLabel={datasetLabel}
       />
